Enable -Wall and -Wsign-compare compiler warnings (#26)
* Initial plan

* Fix -Wsign-compare warnings throughout the codebase

Resolve all -Wsign-compare warnings found when building with
-Wall -Wsign-compare -Wno-unused-parameter -DCMAKE_UNITY_BUILD=OFF.

Strategy used per-site:
- Changed loop counters from GLint/int to GLuint where the variable
  only iterates over unsigned counts (vbo_exec_array.c, api_validate.c,
  context.c, slang_print.c, slang_vartable.c, ss_context.c,
  prog_execute.c).
- Changed GLint min_index/max_index to GLuint in vbo_split_inplace.c
  since they are only assigned from unsigned struct fields.
- Added (GLuint) or (GLint) casts at comparison sites where changing
  the variable type is unsafe or touches public API (shader_api.c,
  slang_codegen.c, slang_compile.c, slang_emit.c, slang_link.c,
  slang_simplify.c, slang_vartable.c, prog_execute.c, texstore.c,
  renderbuffer.c, arbprogparse.c, t_draw.c).
- Cast sizeof/array-size expressions to (GLint) in assert() calls
  in prog_print.c.
- Fixed ternary operator signedness mismatch in slang_emit.c.
- Changed ~0 to ~0u for unsigned dispatch_offset comparisons in
  glapi.c.
- Added (GLint) casts for gl_state_index enum vs int comparisons in
  arbprogparse.c.

No struct field types or function signatures were changed.


* Fix sign-compare warnings and enable -Wall and -Wsign-compare in CMakeLists.txt


* Fix remaining -Wall warnings: unused functions and uninitialized variables


---------

+++ src/shader/prog_print.c
@@ -125,10 +125,10 @@ arb_input_attrib_string(GLint index, GLenum progType)
     };
 
     if (progType == GL_VERTEX_PROGRAM_ARB) {
-	assert(index < sizeof(vertAttribs) / sizeof(vertAttribs[0]));
+	assert(index < (GLint)(sizeof(vertAttribs) / sizeof(vertAttribs[0])));
 	return vertAttribs[index];
     } else {
-	assert(index < sizeof(fragAttribs) / sizeof(fragAttribs[0]));
+	assert(index < (GLint)(sizeof(fragAttribs) / sizeof(fragAttribs[0])));
 	return fragAttribs[index];
     }
 }
@@ -168,10 +168,10 @@ arb_output_attrib_string(GLint index, GLenum progType)
     };
 
     if (progType == GL_VERTEX_PROGRAM_ARB) {
-	assert(index < sizeof(vertResults) / sizeof(vertResults[0]));
+	assert(index < (GLint)(sizeof(vertResults) / sizeof(vertResults[0])));
 	return vertResults[index];
     } else {
-	assert(index < sizeof(fragResults) / sizeof(fragResults[0]));
+	assert(index < (GLint)(sizeof(fragResults) / sizeof(fragResults[0])));
 	return fragResults[index];
     }
 }
